added module navbar.js
