+++ favorites.html
@@ -289,7 +289,7 @@
       </div>
     </footer>
 
-    <script src="/index.js"></script>
+    <script type="module" src="/index.js"></script>
   </body>
 </html>
