+++ index.html
@@ -333,7 +333,7 @@
       </div>
     </footer>
 
-    <script src="/index.js"></script>
+    <script type="module" src="/index.js"></script>
   </body>
 </html>
