+++ public/index.js
@@ -1,5 +1,13 @@
+import { setupNavbar } from "./modules/navbar.js";
+
+document.addEventListener("DOMContentLoaded", () => {
+  setupNavbar();
+});
+
 (function () {
+  // IIFE - Immediately Invoked Function Expression
   function Todo() {
+    // Main function, JS-module wrapped in IIFE to avoid global scope pollution
     const list = document.getElementById("added-list"); // first list (ul) that gets tasks appended
     const favoritesList = document.getElementById("fav_albums"); // second list (ul) with favorite songs
     const searchInput = document.getElementById("search-all-input");
@@ -18,42 +26,6 @@
       return document.createElement("div");
     }
 
-    // Navbar behavior:
-
-    let prevScrollPos = window.scrollY;
-    const navbar = document.getElementById("navigation");
-
-    // get navbar height (it varies depending on resolution):
-    function getNavbarHeight() {
-      return navbar.offsetHeight;
-    }
-
-    window.onscroll = function () {
-      let currentScrollPos = window.scrollY;
-
-      if (prevScrollPos > currentScrollPos) {
-        //  If scrolling up, show navbar:
-        navbar.style.top = "0";
-      } else {
-        // If scrolling down, hide navbar:
-        navbar.style.top = `-${getNavbarHeight()}px`; // Navbar height in pixel - adjustable
-      }
-
-      prevScrollPos = currentScrollPos;
-    };
-
-    // Show navbar when hovered over with mouse:
-    navbar.addEventListener("mouseenter", () => {
-      navbar.style.top = "0";
-    });
-
-    // Hide navbar when mouse leaves navbar area and user scrolls down:
-    navbar.addEventListener("mouseleave", () => {
-      let currentScrollPos = window.scrollY;
-      if (prevScrollPos < currentScrollPos) {
-        navbar.style.top = `-${getNavbarHeight()}px`;
-      }
-    });
 
     ///////////////////////////////////   Themes   ///////////////////////////////////////////
